# api/openapi.yaml
openapi: 3.1.0
info:
  title: bank0 API
  version: 0.1.0
  description: |
    Core-banking backend. Two surfaces, separated by domain:
      * **client** (api.bank0.hnimn.art) — customer-facing money operations
      * **admin**  (portal.bank0.hnimn.art) — operator/back-office operations

    This file is the source of truth. `oapi-codegen` generates the Go
    `ServerInterface` for each tag (client/admin); handler drift is a compile
    error. Edit here, regenerate, fix the compiler.
servers:
  - url: https://api.bank0.hnimn.art
    description: client
  - url: https://portal.bank0.hnimn.art
    description: admin
...
paths:
...
  /transfers/{id}/post:
    post:
      operationId: postTransfer
      tags: [client]
      summary: Post a pending transfer (deferred settlement). Idempotent.
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusResponse" }
  /transfers/{id}/cancel:
    post:
      operationId: cancelTransfer
      tags: [client]
      summary: Cancel a pending transfer (releases the hold). Idempotent.
      parameters: [ { $ref: "#/components/parameters/Id" } ]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReasonRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusResponse" }
...
components:
  securitySchemes:
    # client surface (api.bank0.hnimn.art): JWT bearer from POST /auth/login
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    # admin surface (portal.bank0.hnimn.art): DB-backed cookie session from /login
    sessionCookie:
      type: apiKey
      in: cookie
      name: bank0_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    StatusResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string }
...
    ReasonRequest:
      type: object
      properties:
        reason: { type: string }
...